feat: mp modal improvements
