+++ src/app/register/[resource_id]/components/FormHeader.tsx
@@ -12,17 +12,19 @@ export default function FormHeader({ resource }: FormHeaderProps) {
         <Image
           src="https://edqkxwgbbunlomuzarwt.supabase.co/storage/v1/object/public/assets//HDC-2-mda-logo-05.png"
           alt="Hablemos de cancer"
-          width={150}
-          height={150}
+          width={100}
+          height={100}
           className="object-contain"
           priority
         />
       </div>
       <div>
-        <h1 className="text-2xl xs:text-4xl md:text-md font-semibold text-gray-800 mb-2">
+        <h1 className="text-xl md:text-md font-semibold text-gray-800 mb-2">
           {resource?.name}
         </h1>
       </div>
+
+      <div className="text-sm text-gray-500">{resource?.description}</div>
     </div>
   );
 }
