+++ src/app/register/[resource_id]/components/RegisterFormScreen.tsx
@@ -10,7 +10,7 @@ import {
 import { Form } from "@/components/ui/form";
 import { Resource } from "@/app/schema/resource";
 import SplashLoaderModal from "@/app/register/[resource_id]/components/SplashLoaderModal";
-import MercadoPagoBanner from "@/app/components/MPBanner";
+import MercadoPagoModal from "@/app/components/MercadoPagoModal";
 import FormHeader from "@/app/register/[resource_id]/components/FormHeader";
 import StepProgressBar from "@/app/register/[resource_id]/components/StepProgressBar";
 import FormStepContent from "@/app/register/[resource_id]/components/FormStepContent";
@@ -74,7 +74,7 @@ export default function RegisterFormScreen({
 
   return (
     <div className="flex h-screen flex-col items-center bg-gray-50 p-2 sm:p-4">
-      <div className="w-full max-w-3xl h-full flex flex-col">
+      <div className="w-full max-w-3xl flex flex-col">
         <FormHeader resource={resource} />
 
         <div className="flex-1 flex flex-col shadow-none min-h-0">
@@ -95,10 +95,7 @@ export default function RegisterFormScreen({
                 isLastStep={isLastStep}
               />
 
-              <MercadoPagoBanner
-                amount={resource.price}
-                title={resource.name}
-              />
+              <MercadoPagoModal amount={resource.price} title={resource.name} />
 
               <FormNavigation
                 currentStep={currentStep}
